Moved recording UI to public recording page
Motivations:

 - The purpose of this feature is to let event participants each
   individually create their own recordings. So, a public facing page
   makes much more sense than one in the admin interface.

Actions:

 - Created a public-facing recording interface.
