+++ ui/packages/comhairle/src/routes/(admin)/admin/conversations/[conversation_id]/events/[event_id]/LiveRecorder.svelte
@@ -1,23 +1,7 @@
 <script lang="ts">
-	import { Mic, Square, Loader2, Pause, Play, Trash2 } from 'lucide-svelte';
-	import { Button } from '$lib/components/ui/button';
-	import { Input } from '$lib/components/ui/input';
-	import { WSConnection } from '$lib/api/websockets.svelte';
-	import { notifications } from '$lib/notifications.svelte';
-	import { tryCatchAsync } from '$lib/utils/errorHandling';
-	import { onDestroy, onMount } from 'svelte';
+	import SharedLiveRecorder from '$lib/components/recordings/LiveRecorder.svelte';
 	import type { AudioRecordingDto } from '@crownshy/api-client/api';
 
-	// ── Props ─────────────────────────────────────────────────────────────────
-
-	/**
-	 * Props for LiveRecorder.
-	 *
-	 * @param conversation_id - The parent conversation id.
-	 * @param event_id - The event being recorded under.
-	 * @param recordings - Existing recordings, used to prevent duplicate names.
-	 * @param onComplete - Called after the multipart upload is completed and processing is enqueued.
-	 */
 	type Props = {
 		conversation_id: string;
 		event_id: string;
@@ -26,676 +10,6 @@
 	};
 
 	let { conversation_id, event_id, recordings, onComplete }: Props = $props();
-
-	// ── API response types ────────────────────────────────────────────────────
-	// These mirror the server DTOs and will be replaced by generated client types
-	// once the API spec is published.
-
-	type UploadedPart = {
-		partNumber: number;
-		etag: string;
-		sizeBytes: number;
-	};
-
-	type LiveAudioRecordingDto = {
-		id: string;
-		audioRecordingId: string;
-		multipartUploadId: string;
-		nextPartNumber: number;
-		uploadedParts: UploadedPart[];
-	};
-
-	type CreateLiveAudioRecordingResponse = {
-		recording: AudioRecordingDto;
-		liveAudioRecording: LiveAudioRecordingDto;
-	};
-
-	type PresignLiveAudioRecordingPartResponse = {
-		uploadUrl: string;
-		partNumber: number;
-	};
-
-	type AckLiveAudioRecordingPartResponse = {
-		liveAudioRecording: LiveAudioRecordingDto;
-	};
-
-	type LiveAudioRecordingStateResponse = {
-		liveAudioRecording: LiveAudioRecordingDto;
-	};
-
-	type ProcessRecordingResponse = {
-		message: string;
-		jobId: string;
-	};
-
-	// ── Constants ─────────────────────────────────────────────────────────────
-
-	/** How often (ms) the MediaRecorder emits a `dataavailable` event. */
-	const CHUNK_INTERVAL_MS = 10_000;
-
-	/**
-	 * Minimum accumulated bytes before a multipart part is flushed to S3.
-	 * S3 requires all parts except the final one to be at least 5 MB.
-	 */
-	const MIN_PART_BYTES = 5 * 1024 * 1024;
-
-	// ── Reactive state ────────────────────────────────────────────────────────
-
-	let recordingName = $state('');
-	let phase = $state<'idle' | 'starting' | 'recording' | 'stopping'>('idle');
-	let activeLiveRecordingId = $state<string | null>(null);
-	let liveRecordings = $state<LiveAudioRecordingDto[]>([]);
-
-	const isStarting = $derived(phase === 'starting');
-	const isRecording = $derived(phase === 'recording');
-	const isStopping = $derived(phase === 'stopping');
-	const hasActiveLiveRecording = $derived(activeLiveRecordingId !== null);
-
-	// ── MediaRecorder internals (not reactive — mutated directly) ─────────────
-
-	let mediaRecorder: MediaRecorder | null = null;
-	let mediaStream: MediaStream | null = null;
-
-	/** Sequentially-executed promise chain ensuring chunk uploads do not race. */
-	let uploadChain: Promise<void> = Promise.resolve();
-
-	/** Next expected part number, kept in sync with server ack responses. */
-	let nextPartNumber = 1;
-
-	/** Accumulated blobs waiting to be flushed as a single multipart part. */
-	let pendingBlobs: Blob[] = [];
-	let pendingBytes = 0;
-
-	// ── API helper ────────────────────────────────────────────────────────────
-
-	function liveApiBasePath(): string {
-		return `/api/conversation/${conversation_id}/events/${event_id}/audio_recordings/live`;
-	}
-
-	async function callLiveApi<T>(method: string, path: string, body?: unknown): Promise<T> {
-		const response = await fetch(`${liveApiBasePath()}${path}`, {
-			method,
-			credentials: 'include',
-			headers: body ? { 'content-type': 'application/json' } : undefined,
-			body: body ? JSON.stringify(body) : undefined
-		});
-
-		if (!response.ok) {
-			let message = `${method} ${path} failed (${response.status})`;
-			const parseResult = await tryCatchAsync(() => response.json());
-			if (parseResult.err === null) {
-				if (typeof parseResult.ok?.message === 'string') {
-					message = parseResult.ok.message;
-				} else if (typeof parseResult.ok?.err === 'string') {
-					message = parseResult.ok.err;
-				}
-			}
-			throw new Error(message);
-		}
-
-		return response.json() as Promise<T>;
-	}
-
-	async function loadLiveRecordings(): Promise<void> {
-		const result = await tryCatchAsync(() => callLiveApi<LiveAudioRecordingDto[]>('GET', ''));
-
-		if (result.err !== null) {
-			notifications.send({
-				message: 'Failed to load in-progress live recordings',
-				priority: 'ERROR'
-			});
-			return;
-		}
-
-		liveRecordings = result.ok;
-	}
-
-	function liveRecordingName(liveRecording: LiveAudioRecordingDto): string {
-		return (
-			recordings.find((recording) => recording.id === liveRecording.audioRecordingId)?.name ??
-			'Untitled live recording'
-		);
-	}
-
-	// ── Chunk upload logic ────────────────────────────────────────────────────
-
-	async function putPartToSignedUrl(blob: Blob, url: string): Promise<string> {
-		const response = await fetch(url, { method: 'PUT', body: blob });
-		if (!response.ok) {
-			throw new Error(`Part upload failed (${response.status})`);
-		}
-		const etag = response.headers.get('etag') ?? response.headers.get('ETag') ?? '';
-		return etag.replaceAll('"', '');
-	}
-
-	/**
-	 * Appends a blob to the pending buffer and flushes it as a multipart part when
-	 * the buffer reaches the minimum part size, or when `forceFlush` is true.
-	 *
-	 * Enqueued operations run serially via `uploadChain` so parts are always
-	 * acknowledged in order.
-	 *
-	 * @param liveRecordingId - The id of the active live_audio_recording row.
-	 * @param blob - The audio blob to append, or null to trigger a flush with no new data.
-	 * @param forceFlush - When true the buffer is flushed regardless of size (used at stop time).
-	 */
-	function enqueueChunk(liveRecordingId: string, blob: Blob | null, forceFlush: boolean): void {
-		uploadChain = uploadChain.then(async () => {
-			if (blob && blob.size > 0) {
-				pendingBlobs.push(blob);
-				pendingBytes += blob.size;
-			}
-
-			if (pendingBytes === 0) return;
-			if (!forceFlush && pendingBytes < MIN_PART_BYTES) return;
-
-			const payload = new Blob(pendingBlobs, { type: 'audio/webm' });
-			const partNumber = nextPartNumber;
-			pendingBlobs = [];
-			pendingBytes = 0;
-
-			const presignResult = await tryCatchAsync(() =>
-				callLiveApi<PresignLiveAudioRecordingPartResponse>(
-					'POST',
-					`/${liveRecordingId}/presign`,
-					{ partNumber }
-				)
-			);
-			if (presignResult.err !== null) throw presignResult.err;
-
-			const etag = await putPartToSignedUrl(payload, presignResult.ok.uploadUrl);
-
-			const ackResult = await tryCatchAsync(() =>
-				callLiveApi<AckLiveAudioRecordingPartResponse>('POST', `/${liveRecordingId}/ack`, {
-					partNumber,
-					etag,
-					sizeBytes: payload.size
-				})
-			);
-			if (ackResult.err !== null) throw ackResult.err;
-
-			nextPartNumber = ackResult.ok.liveAudioRecording.nextPartNumber;
-			liveRecordings = liveRecordings.map((liveRecording) =>
-				liveRecording.id === ackResult.ok.liveAudioRecording.id
-					? ackResult.ok.liveAudioRecording
-					: liveRecording
-			);
-		});
-	}
-
-	// ── Volume analysis (Web Audio API) ─────────────────────────────────────────
-
-	/**
-	 * Normalized volume level in the range 0..1, sampled at the rAF rate.
-	 * Drives the scale of the mic icon circle during recording.
-	 */
-	let audioVolume = $state(0);
-
-	let audioContext: AudioContext | null = null;
-	let audioAnalyser: AnalyserNode | null = null;
-	let volumeAnimationFrameId: number | null = null;
-	const recordingWebSocket = new WSConnection();
-	const unsubscribeRecordingWebSocketClose = recordingWebSocket.onClose(() => {
-		if (phase !== 'recording') return;
-
-		void pauseRecording();
-		notifications.send({
-			message: 'Live recording paused because websocket disconnected',
-			priority: 'WARNING'
-		});
-	});
-
-	/**
-	 * Connects the active media stream to an AnalyserNode and starts a
-	 * requestAnimationFrame loop that updates `audioVolume` with the RMS of the
-	 * time-domain signal.
-	 *
-	 * @param stream - The MediaStream that is being recorded.
-	 */
-	function startVolumeAnalysis(stream: MediaStream): void {
-		audioContext = new AudioContext();
-		audioAnalyser = audioContext.createAnalyser();
-		audioAnalyser.fftSize = 256;
-		audioContext.createMediaStreamSource(stream).connect(audioAnalyser);
-
-		const buffer = new Uint8Array(audioAnalyser.frequencyBinCount);
-
-		function tick(): void {
-			if (!audioAnalyser) return;
-			audioAnalyser.getByteTimeDomainData(buffer);
-
-			// Compute RMS of the normalised time-domain deviations (silence = 128).
-			let sum = 0;
-			for (const sample of buffer) {
-				const deviation = (sample - 128) / 128;
-				sum += deviation * deviation;
-			}
-			// Boost by 8 so typical conversational levels map close to 1.
-			audioVolume = Math.min(1, Math.sqrt(sum / buffer.length) * 8);
-			volumeAnimationFrameId = requestAnimationFrame(tick);
-		}
-
-		volumeAnimationFrameId = requestAnimationFrame(tick);
-	}
-
-	/** Cancels the rAF loop and closes the AudioContext. */
-	function stopVolumeAnalysis(): void {
-		if (volumeAnimationFrameId !== null) {
-			cancelAnimationFrame(volumeAnimationFrameId);
-			volumeAnimationFrameId = null;
-		}
-		audioAnalyser = null;
-		void audioContext?.close();
-		audioContext = null;
-		audioVolume = 0;
-	}
-
-	// ── Recorder lifecycle ────────────────────────────────────────────────────
-
-	function releaseMediaDevices(): void {
-		stopVolumeAnalysis();
-		mediaRecorder?.stop();
-		mediaRecorder = null;
-		if (mediaStream) {
-			for (const track of mediaStream.getTracks()) track.stop();
-			mediaStream = null;
-		}
-	}
-
-	/** Requests any buffered audio, stops the recorder, and waits for the upload chain. */
-	async function drainAndStop(liveRecordingId: string): Promise<void> {
-		mediaRecorder?.requestData();
-		mediaRecorder?.stop();
-		// Give the browser a tick to fire the final dataavailable event.
-		await new Promise<void>((resolve) => setTimeout(resolve, 50));
-		releaseMediaDevices();
-
-		// Force-flush any remaining sub-threshold bytes as the final part.
-		enqueueChunk(liveRecordingId, null, true);
-		await uploadChain;
-	}
-
-	// ── Public actions ────────────────────────────────────────────────────────
-
-	async function startRecording(): Promise<void> {
-		const trimmedName = recordingName.trim();
-		if (!trimmedName) {
-			notifications.send({
-				message: 'Enter a name for the recording first',
-				priority: 'ERROR'
-			});
-			return;
-		}
-		if (recordings.some((recording) => recording.name === trimmedName)) {
-			notifications.send({
-				message: `A recording named "${trimmedName}" already exists`,
-				priority: 'ERROR'
-			});
-			return;
-		}
-
-		phase = 'starting';
-
-		const createResult = await tryCatchAsync(() =>
-			callLiveApi<CreateLiveAudioRecordingResponse>('POST', '', {
-				name: trimmedName,
-				fileExtension: 'webm'
-			})
-		);
-
-		if (createResult.err !== null) {
-			const error = createResult.err;
-			notifications.send({
-				message: error instanceof Error ? error.message : 'Failed to start recording',
-				priority: 'ERROR'
-			});
-			phase = 'idle';
-			return;
-		}
-
-		const created = createResult.ok;
-		const liveRecordingId = created.liveAudioRecording.id;
-		activeLiveRecordingId = liveRecordingId;
-		liveRecordings = [...liveRecordings, created.liveAudioRecording];
-
-		nextPartNumber = created.liveAudioRecording.nextPartNumber;
-		pendingBlobs = [];
-		pendingBytes = 0;
-		uploadChain = Promise.resolve();
-
-		const micResult = await tryCatchAsync(() =>
-			navigator.mediaDevices.getUserMedia({ audio: true })
-		);
-
-		if (micResult.err !== null) {
-			const error = micResult.err;
-			notifications.send({
-				message: error instanceof Error ? error.message : 'Microphone access denied',
-				priority: 'ERROR'
-			});
-			// Clean up the DB/S3 row created above.
-			void tryCatchAsync(() => callLiveApi('DELETE', `/${liveRecordingId}`));
-			activeLiveRecordingId = null;
-			phase = 'idle';
-			return;
-		}
-
-		mediaStream = micResult.ok;
-		startVolumeAnalysis(mediaStream);
-
-		const mimeType = MediaRecorder.isTypeSupported('audio/webm;codecs=opus')
-			? 'audio/webm;codecs=opus'
-			: 'audio/webm';
-
-		const recorder = new MediaRecorder(mediaStream, { mimeType });
-		mediaRecorder = recorder;
-
-		recorder.ondataavailable = (event: BlobEvent) => {
-			if (event.data.size > 0) {
-				enqueueChunk(liveRecordingId, event.data, false);
-			}
-		};
-
-		recorder.start(CHUNK_INTERVAL_MS);
-		recordingWebSocket.connect();
-		recordingName = '';
-		phase = 'recording';
-	}
-
-	async function resumeLiveRecording(liveRecordingId: string): Promise<void> {
-		if (phase !== 'idle') {
-			notifications.send({
-				message: 'Pause or stop the current recording before resuming another one',
-				priority: 'WARNING'
-			});
-			return;
-		}
-
-		const liveRecording = liveRecordings.find((recording) => recording.id === liveRecordingId);
-		if (!liveRecording) {
-			notifications.send({
-				message: 'Could not find that live recording',
-				priority: 'ERROR'
-			});
-			return;
-		}
-
-		phase = 'starting';
-		activeLiveRecordingId = liveRecording.id;
-		nextPartNumber = liveRecording.nextPartNumber;
-		pendingBlobs = [];
-		pendingBytes = 0;
-		uploadChain = Promise.resolve();
-
-		const microphoneResult = await tryCatchAsync(() =>
-			navigator.mediaDevices.getUserMedia({ audio: true })
-		);
-
-		if (microphoneResult.err !== null) {
-			activeLiveRecordingId = null;
-			phase = 'idle';
-			notifications.send({
-				message:
-					microphoneResult.err instanceof Error
-						? microphoneResult.err.message
-						: 'Microphone access denied',
-				priority: 'ERROR'
-			});
-			return;
-		}
-
-		mediaStream = microphoneResult.ok;
-		startVolumeAnalysis(mediaStream);
-
-		const mimeType = MediaRecorder.isTypeSupported('audio/webm;codecs=opus')
-			? 'audio/webm;codecs=opus'
-			: 'audio/webm';
-
-		const recorder = new MediaRecorder(mediaStream, { mimeType });
-		mediaRecorder = recorder;
-
-		recorder.ondataavailable = (event: BlobEvent) => {
-			if (event.data.size > 0) {
-				enqueueChunk(liveRecording.id, event.data, false);
-			}
-		};
-
-		recorder.start(CHUNK_INTERVAL_MS);
-		recordingWebSocket.connect();
-		phase = 'recording';
-	}
-
-	async function discardLiveRecording(liveRecordingId: string): Promise<void> {
-		if (activeLiveRecordingId === liveRecordingId) {
-			notifications.send({
-				message: 'Stop or pause this recording before discarding it',
-				priority: 'WARNING'
-			});
-			return;
-		}
-
-		const result = await tryCatchAsync(() => callLiveApi('DELETE', `/${liveRecordingId}`));
-		if (result.err !== null) {
-			notifications.send({
-				message:
-					result.err instanceof Error
-						? result.err.message
-						: 'Failed to discard live recording',
-				priority: 'ERROR'
-			});
-			return;
-		}
-
-		liveRecordings = liveRecordings.filter(
-			(liveRecording) => liveRecording.id !== liveRecordingId
-		);
-	}
-
-	async function pauseRecording(): Promise<void> {
-		const liveRecordingId = activeLiveRecordingId;
-		if (!mediaRecorder || phase !== 'recording' || !liveRecordingId) return;
-
-		phase = 'stopping';
-
-		const pauseResult = await tryCatchAsync(() => drainAndStop(liveRecordingId));
-
-		recordingWebSocket.disconnect();
-
-		if (pauseResult.err !== null) {
-			const error = pauseResult.err;
-			notifications.send({
-				message: error instanceof Error ? error.message : 'Failed to pause recording',
-				priority: 'ERROR'
-			});
-			releaseMediaDevices();
-			activeLiveRecordingId = null;
-			phase = 'idle';
-			return;
-		}
-
-		const deactivateResult = await tryCatchAsync(() =>
-			callLiveApi<LiveAudioRecordingStateResponse>('POST', `/${liveRecordingId}/deactivate`)
-		);
-
-		if (deactivateResult.err !== null) {
-			notifications.send({
-				message:
-					deactivateResult.err instanceof Error
-						? deactivateResult.err.message
-						: 'Failed to release live recording lock',
-				priority: 'WARNING'
-			});
-		}
-
-		activeLiveRecordingId = null;
-		phase = 'idle';
-	}
-
-	async function stopRecording(): Promise<void> {
-		const liveRecordingId = activeLiveRecordingId;
-		if (!liveRecordingId) return;
-
-		phase = 'stopping';
-
-		const stopResult = await tryCatchAsync(() => drainAndStop(liveRecordingId));
-
-		recordingWebSocket.disconnect();
-
-		if (stopResult.err !== null) {
-			const error = stopResult.err;
-			notifications.send({
-				message: error instanceof Error ? error.message : 'Failed to flush recording data',
-				priority: 'ERROR'
-			});
-			releaseMediaDevices();
-			activeLiveRecordingId = null;
-			phase = 'idle';
-			return;
-		}
-
-		const completeResult = await tryCatchAsync(() =>
-			callLiveApi<ProcessRecordingResponse>('POST', `/${liveRecordingId}/complete`)
-		);
-
-		activeLiveRecordingId = null;
-		phase = 'idle';
-		liveRecordings = liveRecordings.filter(
-			(liveRecording) => liveRecording.id !== liveRecordingId
-		);
-
-		if (completeResult.err !== null) {
-			const error = completeResult.err;
-			notifications.send({
-				message: error instanceof Error ? error.message : 'Failed to finalise recording',
-				priority: 'ERROR'
-			});
-			return;
-		}
-
-		notifications.send({
-			message: completeResult.ok.message || 'Recording saved — transcription started',
-			priority: 'SUCCESS'
-		});
-
-		await onComplete?.();
-	}
-
-	// ── Cleanup on component destroy ──────────────────────────────────────────
-
-	onDestroy(() => {
-		unsubscribeRecordingWebSocketClose();
-		recordingWebSocket.disconnect();
-		releaseMediaDevices();
-
-		if (activeLiveRecordingId) {
-			// Fire-and-forget: abort the multipart upload so S3 does not accumulate orphaned parts.
-			void tryCatchAsync(() => callLiveApi('DELETE', `/${activeLiveRecordingId}`));
-		}
-	});
-
-	onMount(() => {
-		void loadLiveRecordings();
-	});
 </script>
 
-<div class="flex flex-col gap-4">
-	{#if liveRecordings.length > 0}
-		<div class="border-border rounded-lg border p-3">
-			<div class="mb-2 text-sm font-semibold">In-progress live recordings</div>
-			<div class="flex flex-col gap-2">
-				{#each liveRecordings as liveRecording (liveRecording.id)}
-					{@const isActiveRow = activeLiveRecordingId === liveRecording.id}
-					<div
-						class="bg-muted/40 flex items-center justify-between gap-3 rounded-md p-2"
-						class:opacity-50={hasActiveLiveRecording && !isActiveRow}
-					>
-						<div class="min-w-0">
-							<div class="truncate text-sm font-medium">
-								{liveRecordingName(liveRecording)}
-							</div>
-							<div class="text-muted-foreground text-sm">
-								Next part: {liveRecording.nextPartNumber}, uploaded: {liveRecording
-									.uploadedParts.length}
-							</div>
-						</div>
-						<div class="flex items-center gap-2">
-							{#if isActiveRow}
-								{#if isStarting}
-									<div
-										class="text-muted-foreground flex items-center gap-2 text-sm"
-									>
-										<Loader2 class="h-4 w-4 animate-spin" />
-										Starting…
-									</div>
-								{:else if isRecording}
-									<div
-										class="bg-destructive/10 flex h-8 w-8 items-center justify-center rounded-full transition-transform duration-75"
-										style="transform: scale({1 + audioVolume * 0.35})"
-										aria-hidden="true"
-									>
-										<Mic class="text-destructive h-4 w-4" />
-									</div>
-									<Button variant="outline" size="sm" onclick={pauseRecording}>
-										<Pause class="mr-2 h-4 w-4" />
-										Pause
-									</Button>
-									<Button variant="destructive" size="sm" onclick={stopRecording}>
-										<Square class="mr-2 h-4 w-4" />
-										Stop &amp; save
-									</Button>
-								{:else if isStopping}
-									<div
-										class="text-muted-foreground flex items-center gap-2 text-sm"
-									>
-										<Loader2 class="h-4 w-4 animate-spin" />
-										Saving…
-									</div>
-								{/if}
-							{:else}
-								<Button
-									variant="outline"
-									size="sm"
-									disabled={hasActiveLiveRecording}
-									onclick={() => resumeLiveRecording(liveRecording.id)}
-								>
-									<Play class="mr-2 h-4 w-4" />
-									Resume
-								</Button>
-								<Button
-									variant="ghost"
-									size="sm"
-									disabled={hasActiveLiveRecording}
-									onclick={() => discardLiveRecording(liveRecording.id)}
-								>
-									<Trash2 class="h-4 w-4" />
-								</Button>
-							{/if}
-						</div>
-					</div>
-				{/each}
-			</div>
-		</div>
-	{/if}
-
-	<div class="flex items-center gap-4" class:opacity-50={hasActiveLiveRecording}>
-		<!--
-			Mic icon acts as the visual anchor for the live recording action.
-			The icon itself is not interactive; the labelled Start button is the trigger.
-		-->
-		<div class="bg-muted flex h-10 w-10 shrink-0 items-center justify-center rounded-full">
-			<Mic class="h-5 w-5" />
-		</div>
-		<Input
-			class="max-w-xs"
-			placeholder="Recording name"
-			disabled={hasActiveLiveRecording}
-			bind:value={recordingName}
-			onkeydown={(e) => {
-				if (e.key === 'Enter') startRecording();
-			}}
-		/>
-		<Button onclick={startRecording} disabled={hasActiveLiveRecording || !recordingName.trim()}
-			>Start recording</Button
-		>
-	</div>
-</div>
+<SharedLiveRecorder {conversation_id} {event_id} {recordings} {onComplete} mode="admin" />
